% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_ts.R
\name{complete_ts}
\alias{complete_ts}
\title{complete_ts}
\usage{
complete_ts(x)
}
\arguments{
\item{x}{data.frame with the columns st_id, month, h_clear and h_cloud}
}
\value{

}
\description{
Fill the missing values between the monthly reference values for radiation via linear interpolation
}
