% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kriging.R
\name{kriging}
\alias{kriging}
\title{kriging}
\usage{
kriging(x, new_data, formula)
}
\arguments{
\item{x}{SpatialPointsDataFrame with observed values}

\item{new_data}{a raster file that defines the points for which predictions are needed}

\item{formula}{a formula, for ordinary kriging use var ~ 1, where var equals the column name in x, where the observations are stored}
}
\value{

}
\description{
Interpolation of point values via ordinary kriging
}
