% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{official_stations}
\alias{official_stations}
\title{Location of the official stations}
\format{
A SpatialPointsDataFrame with 15 features and 4 variables:
\describe{
\item{st_id}{id of the station}
\item{elev}{elevation of the station in meters above sea level}
\item{slope}{slope of the station in degrees}
\item{aspect}{aspect of the station in degrees}
}
}
\usage{
official_stations
}
\description{
A SpatialPointsDataFrame with the location of the official stations from the national meteorological service
}
\keyword{datasets}
