% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rad_longterm}
\alias{rad_longterm}
\title{Table with daily measurements of solar irradiation from the official stations}
\format{
A data frame with 35430 rows and 4 variables:
\describe{
\item{st_id}{id of the official stations}
\item{date}{measurement date}
\item{month}{measurement month}
\item{irradiation}{daily solar irradiation in kWh/m2}
}
}
\usage{
rad_longterm
}
\description{
A data.frame containing measured solar irradiation from the official stations with a long timeseries of measurements
}
\keyword{datasets}
