% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timeseries}
\alias{timeseries}
\title{Table with daily measurements from the official stations.}
\format{
A data frame with 16425 rows and 4 variables:
\describe{
\item{st_id}{id of the official stations}
\item{date}{measurement date}
\item{tmean}{daily mean temperature in deg C}
\item{irradiation}{daily solar irradiation in kWh/m2}
}
}
\usage{
timeseries
}
\description{
A data.frame with measurements from the official stations
}
\keyword{datasets}
