% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation_phenology}
\alias{validation_phenology}
\title{Table with validation data from the phenological surveys.}
\format{
A data frame with 210 rows and 7 variables:
\describe{
\item{st_id}{id of the validation stations}
\item{year}{year}
\item{stage}{phenological stage}
\item{survey}{starting day of year for the phenological stage from the vineyard survey}
\item{measured}{modeled starting day of year for the phenological stage from the measured temperature from the validation station}
\item{lapse_rate}{modeled starting day of year for the phenological stage from the predictions of the lapse-rate model}
\item{topoclimate}{modeled starting day of year for the phenological stage from the predictions of the topoclimate model}
}
}
\usage{
validation_phenology
}
\description{
A data.frame containing a column for station id, year, phenological stage, phenological timing from the vineyard surveys, modeled phenological timing from measured temperature as well as from the lapse-rate and topoclimate model predictions
}
\keyword{datasets}
