% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation_stations}
\alias{validation_stations}
\title{Location of the validation stations}
\format{
A SpatialPointsDataFrame with 15 features and 4 variables:
\describe{
\item{st_id}{id of the validation stations}
\item{elev}{elevation of the station in meters above sea level}
\item{slope}{slope of the station in degrees}
\item{aspect}{aspect of the station in degrees}
\item{aspect_c}{aspect classification among east, west, south or north}
\item{elev_c}{elevation class of the station}
}
}
\usage{
validation_stations
}
\description{
A SpatialPointsDataFrame with the location of the validation stations that are placed within vineyards throughout the study area
}
\keyword{datasets}
