% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation_tair}
\alias{validation_tair}
\title{Table with validation data from the validation stations.}
\format{
A data frame with 28470 rows and 5 variables:
\describe{
\item{st_id}{id of the validation stations}
\item{date}{measurement date}
\item{measured}{measured air temperature from the validation stations}
\item{lapse_rate}{modeled air temperature from the lapse-rate model}
\item{topoclimate}{modeled air temperature from the topoclimate model}
}
}
\usage{
validation_tair
}
\description{
A data.frame containing a column for station id, date and measured air temperature from the validation stations as well as predictions from the lapse-rate and topoclimate models
}
\keyword{datasets}
