% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rad_longterm}
\alias{rad_longterm}
\title{Table with daily observations of solar irradiation for a long time period.}
\format{
A data frame with 16425 rows and 4 variables:
\describe{
\item{st_id}{id of the official stations}
\item{date}{measurement date}
\item{month}{measurement month}
\item{irradiation}{daily solar irradiation in kWh/m2}
}
}
\usage{
rad_longterm
}
\description{
A data.frame containing a column for station id, date and daily solar irradiation
}
\keyword{datasets}
